Add custom Bootstrap popover

This function creates a Bootstrap popover containing additional information to display alongside a given text.

Usage

with_popover(text, title = "More Info", content, placement = "right")

Arguments

text

The text to display.

title

The title of the popover. Defaults to "More Info".

content

The content to display inside the popover.

placement

The placement of the popover relative to the text. Defaults to "right".

Value

HTML Bootstrap popover containing the specified content.

Examples

if (FALSE) { # \dontrun{
with_popover(
  text = "Some Example Text",
  content = shiny::tagList(
    shiny::span("First description line."),
    shiny::br(),
    shiny::span("Second description line.")
  )
)
} # }